add new file for solution long code
